move flask route handlers into routes module
